appserver/pyproject.toml
[project]
name = "lifelike-appserver"
version = "0.0.0"
requires-python = "==3.10.*"
dependencies = [
  "attrs==26.1.0",
  "bcrypt==3.2.0",
  "flask-marshmallow==1.4.0",
  "flask==3.1.3",
  "flask-caching==2.3.1",
  "flask-cors==6.0.0",
  "flask-httpauth==4.8.0",
  "flask-migrate==2.5.3",
  "flask-sqlalchemy>=3.0,<3.1",
  "gunicorn==22.0.0",
  "marshmallow-sqlalchemy==1.5.0",
  "openpyxl==3.0.9",
  "pyjwt==2.12.1",
  "psycopg2-binary==2.9.11",
  "bioc==2.1",
  "pdfminer-six==20260107",
  "graphviz==0.19.1",
  "pypdf4==1.27.0",
  "sqlalchemy-searchable==2.2.0",
  "elasticsearch==9.3.0",
  "IPy==1.01",
  "sentry-sdk[flask]==2.58.0",
  "redis==4.0.2",
  "flask-apispec==0.11.4",
  "requests==2.33.1",
  "pandas==2.3.3",
  "xlsxwriter==3.2.9",
  "python-json-logger==4.0.0",
  "intervaltree==3.2.1",
  "paramiko==3.4.0",
  "google-cloud-storage==3.10.1",
  "fastjsonschema==2.21.2",
  "marshmallow-dataclass==8.7.1",
  "marshmallow-enum==1.5.1",
  "deepdiff==8.6.2",
  "timeflake==0.4.3",
  "scipy==1.15.3",
  "sqlalchemy==1.4.54",
  "azure-storage-file==2.1.0",
  "azure-storage-blob==12.28.0",
  "neo4j==6.1.0",
  "python-magic==0.4.27",
  "sendgrid==6.12.5",
  "werkzeug==3.1.6",
  "markupsafe==2.1.5",
  "lmdb==1.2.1",
  "pillow==12.2.0",
  "pyparsing==2.4.7",
  "svg-stack==0.1.0",
  "elastic-apm[flask]",
  "pyyaml==6.0.1",
]

[dependency-groups]
dev = [
  "ipython==8.10.0",
  "pytest==9.0.3",
  "ruff==0.15.10",
  "pycodestyle==2.14.0",
  "mypy==1.19.1",
  "httpretty==1.1.4",
  "types-paramiko==2.8.5",
  "types-redis==4.6.0.20241004",
  "types-requests==2.33.0.20260402",
  "pytest-cov==7.0.0",
]
